feat: 컴포넌트 분리 및 즐겨찾기 버튼 모두 추가
