+++ src/components/godplaces/map/KoreanRestaurantMarkers.tsx
@@ -1,0 +1,28 @@
+import { useGodplacesStore } from '@/stores/godplacesStore';
+import { getGodplacePositions } from '@/lib/helper/getGodplacePositions';
+import CustomMarker from './CustomMarker';
+
+export default function KoreanRestaurantMarkers() {
+  const godplaces = useGodplacesStore((state) => state.godplaces);
+
+  const koreanRestaurantPositions = getGodplacePositions(
+    godplaces,
+    true,
+    '한식',
+  );
+  const koreanRestaurantMarkerOrigin = { x: 16, y: 0 };
+
+  return (
+    <>
+      {koreanRestaurantPositions &&
+        koreanRestaurantPositions.map((position) => (
+          <div key={`krRestaurant-${position.lat},${position.lng}`}>
+            <CustomMarker
+              position={position}
+              markerOrigin={koreanRestaurantMarkerOrigin}
+            />
+          </div>
+        ))}
+    </>
+  );
+}
